+++ DSAObjectWeaponHandAndLongRangeWeapon.m
@@ -31,8 +31,8 @@
                 inSubCategory: (NSString *) subCategory
              inSubSubCategory: (NSString *) subSubCategory
                    withWeight: (NSInteger) weight
-                    withPrice: (NSInteger) price
-                   withLength: (NSInteger) length
+                    withPrice: (float) price
+                   withLength: (float) length
                 withHitPoints: (NSArray *) hitPoints
               withHitPointsKK: (NSInteger) hitPointsKK
               withBreakFactor: (NSInteger) breakFactor
